# docs/swagger.yaml
basePath: /
definitions:
...
  dto.Pagination:
    properties:
      pageNum:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  echox.PageInfo:
    properties:
      pageNum:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  echox.Response:
    properties:
      code:
        type: string
      data: {}
      message: {}
      page:
        $ref: '#/definitions/echox.PageInfo'
    type: object
...
  system.DictForm:
    properties:
      dictCode:
        maxLength: 100
        type: string
      id:
        type: integer
      name:
        maxLength: 100
        type: string
      remark:
        maxLength: 255
        type: string
      status:
        type: integer
    required:
    - dictCode
    - name
    type: object
  system.DictItem:
    properties:
      createBy:
        type: integer
      createTime:
        type: string
      dictCode:
        type: string
      id:
        type: integer
      isDeleted:
        type: integer
      label:
        type: string
      remark:
        type: string
      sort:
        type: integer
      status:
        type: integer
      tagType:
        type: string
      updateBy:
        type: integer
      updateTime:
        type: string
      value:
        type: string
    type: object
  system.DictItemForm:
    properties:
      dictCode:
        type: string
      id:
        type: integer
      label:
        maxLength: 100
        type: string
      remark:
        maxLength: 255
        type: string
      sort:
        type: integer
      status:
        type: integer
      tagType:
        maxLength: 50
        type: string
      value:
        maxLength: 100
        type: string
    required:
    - label
    - value
    type: object
  system.DictItemOptionVO:
    properties:
      label:
        type: string
      tagType:
        type: string
      value:
        type: string
    type: object
  system.DictItemQueryResult:
    properties:
      list:
        items:
          $ref: '#/definitions/system.DictItem'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
  system.DictPageVO:
    properties:
      createTime:
        type: string
      dictCode:
        type: string
      id:
        type: integer
      name:
        type: string
      remark:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:2222
info:
  contact:
    email: liusha@email.cn
    name: LiuSha
  description: Echo Admin 后台管理系统 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/top-system/light-admin
  title: Echo Admin API
  version: "1.0"
paths:
...
  /api/v1/dicts:
    get:
      parameters:
      - description: 关键字
        in: query
        name: keywords
        type: string
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/system.DictPageVO'
                  type: array
              type: object
      summary: 字典分页列表
      tags:
      - Dict
    post:
      parameters:
      - description: 字典信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.DictForm'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
      summary: 新增字典
      tags:
      - Dict
  /api/v1/dicts/{dictCode}/items:
    get:
      parameters:
      - description: 字典编码
        in: path
        name: dictCode
        required: true
        type: string
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  $ref: '#/definitions/system.DictItemQueryResult'
              type: object
      summary: 字典项列表
      tags:
      - Dict
    post:
      parameters:
      - description: 字典编码
        in: path
        name: dictCode
        required: true
        type: string
      - description: 字典项信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.DictItemForm'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
      summary: 新增字典项
      tags:
      - Dict
  /api/v1/dicts/{dictCode}/items/{itemId}:
    put:
      parameters:
      - description: 字典编码
        in: path
        name: dictCode
        required: true
        type: string
      - description: 字典项ID
        in: path
        name: itemId
        required: true
        type: integer
      - description: 字典项信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.DictItemForm'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
      summary: 修改字典项
      tags:
      - Dict
  /api/v1/dicts/{dictCode}/items/{itemId}/form:
    get:
      parameters:
      - description: 字典编码
        in: path
        name: dictCode
        required: true
        type: string
      - description: 字典项ID
        in: path
        name: itemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  $ref: '#/definitions/system.DictItemForm'
              type: object
      summary: 获取字典项表单数据
      tags:
      - Dict
  /api/v1/dicts/{dictCode}/items/{itemIds}:
    delete:
      parameters:
      - description: 字典编码
        in: path
        name: dictCode
        required: true
        type: string
      - description: 字典项ID，多个以英文逗号分割
        in: path
        name: itemIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
      summary: 删除字典项
      tags:
      - Dict
  /api/v1/dicts/{dictCode}/items/options:
    get:
      parameters:
      - description: 字典编码
        in: path
        name: dictCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/system.DictItemOptionVO'
                  type: array
              type: object
      summary: 字典项下拉选项
      tags:
      - Dict
  /api/v1/dicts/{id}:
    put:
      parameters:
      - description: 字典ID
        in: path
        name: id
        required: true
        type: integer
      - description: 字典信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.DictForm'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
      summary: 修改字典
      tags:
      - Dict
  /api/v1/dicts/{id}/form:
    get:
      parameters:
      - description: 字典ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/echox.Response'
            - properties:
                data:
                  $ref: '#/definitions/system.DictForm'
              type: object
      summary: 获取字典表单数据
      tags:
      - Dict
  /api/v1/dicts/{ids}:
    delete:
      parameters:
      - description: 字典ID，多个以英文逗号分割
        in: path
        name: ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/echox.Response'
      summary: 删除字典
      tags:
      - Dict
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"